+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -1072,10 +1072,12 @@ class RanaBrowser(QWidget):
 
     @pyqtSlot()
     def enable(self):
+        self.breadcrumbs_stack.currentWidget().setEnabled(True)
         self.rana_browser.setEnabled(True)
 
     @pyqtSlot()
     def disable(self):
+        self.breadcrumbs_stack.currentWidget().setEnabled(False)
         self.rana_browser.setEnabled(False)
 
     def on_project_tab_changed(self, index):
@@ -1106,6 +1108,14 @@ class RanaBrowser(QWidget):
         if current_widget and hasattr(current_widget, "refresh"):
             current_widget.refresh()
             self.last_refresh_time = time.time()
+
+    def reset(self):
+        self.disable()
+        self.rana_browser.setCurrentIndex(0)
+        for widget_idx in range(self.breadcrumbs_stack.count()):
+            self.breadcrumbs_stack.widget(widget_idx).back_to_root()
+        self.projects_browser.refresh()
+        self.enable()
 
     @pyqtSlot()
     def refresh(self):